requirements.txt
pyjokes
requests
beautifulsoup4
speechrecognition
mutagen
ibm-watson
opencv-python
beeply
pytube
pafy
numpy
pillow
pyautogui
